refactor(codegen): kebab-case plugin folder names
