+++ scripts/codegen/src/plugins/default-children/schema.ts
