+++ scripts/codegen/src/plugins/html-attrs/index.ts
