+++ scripts/codegen/src/plugins/imperative-props/check.ts
