+++ scripts/codegen/src/plugins/imperative-props/schema.ts
