+++ scripts/codegen/src/plugins/void-elements/analyze.ts
