+++ scripts/codegen/src/plugins/void-elements/check.test.ts
