+++ scripts/codegen/src/plugins/void-elements/check.ts
