+++ scripts/codegen/src/plugins/void-elements/index.ts
